Fix Pierre's issue in Asana by copying metrics
Metrics are now copied from the monospace font into the newly created
font. This may not be the cleanest solution, but it ensures that the
monospacified font does not cause the line height to change. A manual
fix for Asana Math (the font this was originally reported for) is to
copy the Typographic OS/2 metrics into the HHEA metrics. Following
FontForge's recommendations (setting the HHEA metrics to the Win Ascent
and Descent) does not work (it makes the line height huge, which isn't
visible in Emacs 25+ (it ignores it following a bug report of a few
months ago, but it is visible and annoying in Emacs ≤ 24, and LibreOffice)