Fix glyphs with references (2)
[why]
The previous fix can distort the composite glyphs if the width change is
drastic.

[how]
It is rather hard to adjust all the layers correctly.
The simple approach is to replace the composit glyphs with their actual
component-outlines. After that the scaling can happen as any ordinary
glyph.

The drawback is that the font needs more memory, especially if it has a
large amount of composite glyphs.

We need to do this in two loops because the unlinking must happen before
any referenced glyph has been modified.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>